VDecimate: Fix decimation of the first cycle
The first frame's metrics are always zero, thus the first frame is
always considered a duplicate, which means it's always dropped. Avoid
this by copying the second frame's maximum block difference.

Also make the first frame's total difference just a bit higher than the
scene change threshold, so it's always considered a scene change. This
way the first frame will be dropped when there are no duplicates and no
other scene changes in the cycle.