VDecimate: Use a much smaller cache for the metrics
If the "ovr" parameter is not used, the length of the clip no longer
influences the memory usage. Clips with lengths approaching the
maximum can be processed without running out of memory.

If the "ovr" parameter is used, VDecimate will still allocate one byte
for every cycle. If the clip has the maximum length allowed (2**31-1),
that's 409 MiB in the typical case (cycle=5), and 1 GiB in the worst
case (cycle=2).

The metrics cache uses about 6 KiB.

The speed is unchanged when requesting frames in order.

Fixes #203.