sub.TextFile: Add support for more formats
Convert SRT, Webvtt, Microdvd, etc. to ASS on the fly with ffmpeg.

The explicit dependency on iconv also ensures that the "charset"
parameter is never ignored. Previously the "charset" parameter relied on
libass being compiled with iconv support, which is impossible to check
for and not at all a given.