core: fix wrapping for nfMakeLinear source filters
The previous wrapper implementation only wraps the getFrame function.
However, as it also overrides instanceData as well, when the filter's
free function is called, it is guaranteed to receive the incorrect
instanceData, leading to SIGSEGV when freeing the filter.

We fix this by really implementing a wrapper for nfMakeLinear source
filters: it wraps both getFrame and free functions.

Additionally, we also fix another subtle bug of the previous implementation:
it was calling the filter's free with VSAPI4 instead of VSAPI3.

Signed-off-by: akarin <i@akarin.info>