get_frame(_async_raw): Remove gc.collect from every call.
Garbage collections are slow and should be used sparingly.
As the Python documentation writes:

> With no arguments, run a full collection.

A full collection means "stop the world" on all threads currently
used by the python interpreter. This is slow, especially if run
in a get_frame_async-call.

The RawNode.frames call runs these frames in bulk and only calls
them at the end.